knife file prelim moves - see msg for git log info
Preliminary/unusual moves that we can't just do with the wildcard
that's coming up in the next commit.

To see the complete history for these files, use 'git log --follow',
or use 'git config log.follow true'  to make it the default behavior in
this repository; or 'git config --globa log.follow true' to make it the
global default.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>